+++ src/Aardvark.Data.Collada/RELEASE_NOTES.md
@@ -1,6 +1,10 @@
+### 5.2.10
+- Added PerfectReflection to ColladaMaterial
+- Fixed ColladaMaterial diffuse/specular fallback values
+
 ### 5.2.9
 - Updated to NET 8 and Aardvark.Base 5.3
-- 
+ 
 ### 5.2.9-prerelease0001
 - Updated to NET 8 and Aardvark.Base 5.3 (prerelease)